remove unused pump rollup components
