--- src/app/report-rollup/psat-rollup/psat-rollup-graphs/psat-rollup-graphs.component.css
@@ -1,10 +1,0 @@
-@media print {
-    .print-rollup-pie-container {
-        padding-bottom: 35px;
-        padding-left: 34px;
-    }
-}
-
-.pad-top {
-    padding-top: 20px;
-}
