--- src/app/report-rollup/psat-rollup/psat-rollup-graphs/psat-rollup-graphs.component.spec.ts
@@ -1,25 +1,0 @@
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { PsatRollupGraphsComponent } from './psat-rollup-graphs.component';
-
-describe('PsatRollupGraphsComponent', () => {
-  let component: PsatRollupGraphsComponent;
-  let fixture: ComponentFixture<PsatRollupGraphsComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ PsatRollupGraphsComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(PsatRollupGraphsComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
