+++ src/app/report-rollup/psat-rollup/psat-rollup-print/psat-rollup-print.component.html
@@ -1,4 +1,4 @@
-<div class="print-break-before show-print" *ngIf="settings">
+<!-- <div class="print-break-before show-print" *ngIf="settings">
 
     <div class="row justify-content-center row-bottom-divider">
         <div class="col-12">
@@ -8,7 +8,7 @@
     <!-- <div class="row justify-content-center print-avoid-break" *ngFor="let calculator of calculators">
         <app-pre-assessment-print [calculator]="calculator" [settings]="settings">
         </app-pre-assessment-print>
-    </div> -->
+    </div> --
     <div class="print-avoid-break">
         <div class="row justify-content-center">
             <div class="col-12 header text-center">
@@ -33,4 +33,4 @@
             </div>
         </div>
     </div>
-</div>
+</div> -->
