--- src/app/report-rollup/psat-rollup/psat-rollup-pump-summary-table/psat-rollup-pump-summary-table.component.html
@@ -1,2 +1,0 @@
-<app-rollup-summary-table [tableData]="tableData" [equipmentType]="'Pump'" [energyUnit]="settings.powerMeasurement">
-</app-rollup-summary-table>
