--- src/app/report-rollup/psat-rollup/psat-rollup-pump-summary/psat-rollup-pump-summary.component.css
@@ -1,42 +1,0 @@
-.table-row{
-    padding-top: 14px;
-}
-
-.graph-select{
-    width: 100%;
-}
-
-
-tr > td,
-tr > th{
-    text-align: center;
-}
-
-tr > td {
-    vertical-align: middle;
-}
-
-
-.name{
-    border-left: solid 10px;
-    padding-left: 4px;
-}
-
-th{
-    width:9%;
-}
-
-.table-striped tbody tr:nth-of-type(odd) {
-    background-color: rgba(0, 0, 0, 0.12);
-}
-
-@media print {
-    tr > td,
-    tr > th {
-        font-size: 18px !important;
-    }
-
-    .print-section {
-        page-break-inside: avoid;
-    }
-}
